requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.18
httpx==0.28.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
bcrypt==4.0.1
email-validator==2.2.0
psycopg[binary]==3.2.3
certifi==2024.8.30
